put RawString into separate module
