+++ crates/hcl-edit/src/lib.rs
@@ -9,13 +9,22 @@ pub(crate) mod encode;
 pub mod error;
 pub mod expr;
 pub mod parser;
+mod raw_string;
 pub mod repr;
 pub mod structure;
 pub mod template;
 mod util;
 
-pub use self::error::Error;
+#[doc(inline)]
+pub use self::raw_string::RawString;
 
-// Re-exported for convenience;
+use self::error::Error;
+
+/// Core concepts available for glob import.
+pub mod prelude {
+    pub use crate::repr::{Decorate, Span};
+}
+
+// Re-exported for convenience.
 #[doc(inline)]
 pub use hcl_primitives::{Ident, InternalString, Number};
