+++ crates/hcl-edit/src/parser/mod.rs
@@ -13,10 +13,10 @@ use self::expr::expr;
 use self::structure::body;
 use self::template::template;
 use crate::expr::Expression;
-use crate::repr::{Decorate, Decorated, Despan, RawString, SetSpan};
+use crate::repr::{Decorate, Decorated, Despan, SetSpan};
 use crate::structure::Body;
 use crate::template::Template;
-use crate::{Ident, InternalString, Number};
+use crate::{Ident, InternalString, Number, RawString};
 use hcl_primitives::ident;
 use std::borrow::Cow;
 use std::str::FromStr;
