+++ crates/hcl-edit/src/template/mod.rs
@@ -1,8 +1,8 @@
 use crate::encode::{Encode, EncodeState};
 use crate::expr::Expression;
-use crate::repr::{Decor, Decorate, Decorated, Despan, RawString, SetSpan, Span, Spanned};
+use crate::repr::{Decor, Decorate, Decorated, Despan, SetSpan, Span, Spanned};
 use crate::util::{dedent_by, min_leading_whitespace};
-use crate::{Ident, InternalString};
+use crate::{Ident, InternalString, RawString};
 use std::fmt;
 use std::ops::Range;
